Split ops tests
