+++ crates/core_simd/tests/f32_ops.rs
@@ -1,0 +1,4 @@
+#[macro_use]
+#[path = "ops_macros.rs"]
+mod macros;
+impl_float_tests! { SimdF32, f32, i32 }
