+++ crates/core_simd/tests/i32_ops.rs
@@ -1,0 +1,4 @@
+#[macro_use]
+#[path = "ops_macros.rs"]
+mod macros;
+impl_signed_tests! { SimdI32, i32 }
