+++ crates/core_simd/tests/u64_ops.rs
@@ -1,0 +1,4 @@
+#[macro_use]
+#[path = "ops_macros.rs"]
+mod macros;
+impl_unsigned_tests! { SimdU64, u64 }
